move speak to onEnterFocusGroup
