+++ packages/story-editor/src/components/canvas/framesLayer.js
@@ -23,14 +23,15 @@ import { memo, useRef, useCallback } from '@googleforcreators/react';
 import { __ } from '@googleforcreators/i18n';
 import { PAGE_WIDTH } from '@googleforcreators/units';
 import { STORY_ANIMATION_STATE } from '@googleforcreators/animation';
+import {
+  themeHelpers,
+  useKeyDownEffect,
+  useLiveRegion,
+} from '@googleforcreators/design-system';
 
 /**
  * Internal dependencies
  */
-import {
-  themeHelpers,
-  useKeyDownEffect,
-} from '@googleforcreators/design-system';
 import { DESIGN_SPACE_MARGIN, STABLE_ARRAY } from '../../constants';
 import {
   useStory,
@@ -65,7 +66,12 @@ const FocusContainer = styled.div`
 `;
 
 const FOCUS_CONTAINER_MESSAGE = __(
-  'Canvas Area. To navigate into the page, press Enter. Press Tab to move to the group or element.',
+  'Canvas Area. To navigate into the page, press Enter.',
+  'web-stories'
+);
+
+const FRAME_ELEMENT_MESSAGE = __(
+  'To exit the canvas area, press Escape. Press Tab to move to the next group or element.',
   'web-stories'
 );
 
@@ -180,6 +186,8 @@ function FrameElements() {
 
 function FramesLayer() {
   const canvasRef = useRef();
+  const speak = useLiveRegion();
+
   const enterFocusGroup = useEditLayerFocusManager(
     ({ enterFocusGroup }) => enterFocusGroup
   );
@@ -199,8 +207,9 @@ function FramesLayer() {
         groupId: FOCUS_GROUPS.ELEMENT_SELECTION,
         cleanup: () => canvasRef.current?.focus(),
       });
+      speak(FRAME_ELEMENT_MESSAGE);
     },
-    [enterFocusGroup]
+    [enterFocusGroup, speak]
   );
 
   return (
